requirements.txt
pytorch_lightning==1.9.0
Pillow==10.0.0
opencv-python==4.9.0.80
transformers==4.37.0
taming-transformers-rom1504==0.0.6
tqdm==4.66.1
numpy==1.24.4
kornia==0.7.1
webdataset==0.2.86
omegaconf==2.2.3
einops==0.7.0
scikit-image==0.21.0
pyMCubes==0.1.4
tensorboard==2.14.0
six==1.16.0
trimesh
easydict
ftfy
regex
matplotlib
imageio
imageio-ffmpeg
scipy
pyransac3d
torch_efficient_distloss
nerfacc==0.3.3
rich
rembg==2.0.53
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
git+https://github.com/openai/CLIP.git
